PlayerManager, PlayerActions, Player prefab, and InputSystem package
Added a PlayerManager that can move and turn using the joysticks on a controller. Left joystick moves the character, right joystick turns it. Tried to design the script so it can easily be scaled in the future. Also added a package called InputSystem that's used to handle all of the inputs from the controller. All online sources I found recommended using this, so it's what I went with. Also added the Player prefab, but the texture is just temporary so we can visually see it.