Interactions
Added systems to let the user interact with objects. The InteractionText can go on any object and, as long as the object has a collider, it allows the player to read any preset text on that object. Also added a Text tag that must go on those objects. Also added an InteractionZone script which deals with the collider that checks in front of the player to see if any objects are in front of them.