Dialogue, Animation Template, and Changing Control Maps
Added the ability to change control maps using the InputManager object. InputManager has a function in it's script called swapMap(string) that lets you select which map to use based on its name, and it returns true or false based on whether or not it successfully switched.
Also added an Animation template to PlayerManager which will let us easily implement the animations once they are ready.
Also implemented the dialogue into the Player which will allow us to easily set up interactions with anything that the player will be able to read. In order to set that up all you need to do is put the DialogueTrigger script onto an object that has a collider and the Text tag.