Modified Ice Physics
Modified Chase's ice physics to be a bit cleaner. When the player enters the ice, a new variable in PlayerManager called movementType gets set to 1. This allows the user to persistently move in the direction they were going when they stepped onto the ice. When an ice block stops sliding with the player, the movementType is set to 2 which allows the player a burst of speed in one direction before movementType is set back to 1.

I completely overhauled the Iceblock script to just change it's velocity to one set direction based on whatever direction was receiving more momentum.