LoadZone
Added a LoadZone prefab which can be used to move from scene to scene. To use it, you just need to place it on your scene, give it the name of the scene it's going to, and the x and y coordinates of where the player needs to be in that new scene. You also will likely need to rescale it's collider.

Also added a GameManager. For now it just keeps track of the location of where the player is moving to in the new scene.