Player Hitboxes, DungeonCave, Animations
Worked on the player hitboxes and changed where the collider scripts occur.
Worked on the DungeonCave and added more functionality to the falling script.
Worked on animations to rotate where the player wants to look, and added keyboard arrow keys to rotate the player.
Animations for rotating while standing still currently do not work, and need to be altered.