Combat
Added the first player combat mechanic allowing the player to swing a sword. Code is in place for checking the inventory to see what item is actually being used, but that doesn't work yet. The sword has a collider that will be used to damage enemies. You can attack with ZR or space. The current texture in place for the sword is a placeholder, Matt's texture will replace it soon.