Made Prowler Enemies
Made the Prowler Enemies, that will try and ruthless hunt down the player. they are faster then the player and they have a massive jump that will pin the player to a wall. the prowler also have a very fast but weak attacks.  also made some changes to the slime, made it a little slower.